verify_dist_output/requirements.txt
selenium
pytest
allure-pytest
webdriver-manager
pandas
openpyxl